add random column selection

+++ GLP/Classes/Utilities/SLCrossValidation.h
@@ -194,7 +194,7 @@ SLCrossValidationResults SLCrossValidation<T>::crossValidation(const MatrixXd& X
     if ( useShuffledData )
     {
 //#pragma omp parallel for
-        for (int i=0; i<X.rows(); ++i)
+        for (ssize_t i=0; i<X.rows(); ++i)
         {
             shuffledX.row(i) = X.row(randomIndexs[i]);
             shuffledY.row(i) = Y.row(randomIndexs[i]);
